Fix default compilation flags
The -ansi flag seems to be a gcc feature, and doesn't work with other
compilers such as clang; a more portable way to write this seems to be
-std=c90

The -std=c90 and -pedantic flags are now mandatory, since code should
always compile with them. On the other hand -Werror and -Wfatal-errors
are mostly for ease of use.