Copy history file path to internal buffer
The history file path used to be stored as a pointer to the string
passed by the user. This is not a good idea, since we don't know what
the user will do with the string in subsequent calls to ll_read. It is
also such a rookie mistake I don't even know why I'm explaining myself.

There's an interesting issue with strdup(), or lack of such function
with the standard configuration. Instead, the expression
strcpy(malloc(strlen(foo) + 1, foo) is used. A little cumbersome, but
it works.

Introduces a potential issue because the history path is never
deallocated. This can be fixed by either using a static biffer for the
path or having a cleanup function for the littleline library.